+++ src/main/java/com/derniamepoirier/CardGeneration/Card.java
@@ -2,6 +2,7 @@ package com.derniamepoirier.CardGeneration;
 
 
 import com.derniamepoirier.Utils.DatastoreGetter;
+import com.derniamepoirier.Utils.PaintImageUtils;
 import com.google.appengine.api.datastore.*;
 import com.google.appengine.api.images.Image;
 import com.google.appengine.api.images.*;
@@ -51,12 +52,48 @@ public class Card {
     private double probability;
 
     /**
+     * Constructor of {@link Card}
+     *
+     * @param id                id of the card
+     * @param tags              Tags of the card
+     * @param pixabayPageURL    URL of the image page
+     * @param pixabayImageURL   URL to download Pixabay image
+     * @param pixabayAuthorName Author of the original image
+     * @param probability       probability between 0 and 1 to draw the card
+     * @param cardImageURL      url of the card image
+     */
+    public Card(long id, String[] tags, URL pixabayPageURL, URL pixabayImageURL, URL cardImageURL, String pixabayAuthorName, double probability) {
+        this(id, tags, pixabayPageURL, pixabayImageURL, pixabayAuthorName);
+        this.probability = probability;
+        this.cardImageURL = cardImageURL;
+    }
+
+    /**
+     * Constructor of {@link Card}
+     *
+     * @param id                id of the card
+     * @param tags              Tags of the card
+     * @param pixabayPageURL    URL of the image page
+     * @param pixabayImageURL   URL to download Pixabay image
+     * @param pixabayAuthorName Author of the original image
+     */
+    private Card(long id, String[] tags, URL pixabayPageURL, URL pixabayImageURL, String pixabayAuthorName) {
+        this.id = id;
+        this.tags = tags;
+        this.pixabayPageURL = pixabayPageURL;
+        this.pixabayImageURL = pixabayImageURL;
+        this.pixabayAuthorName = pixabayAuthorName;
+        this.probability = Math.random();
+    }
+
+    /**
      * Generate array of {@link Card} from JSON Pixabay API response
      * If the card already exists in {@link DatastoreService}, the {@link Card} won't be returned
+     *
      * @param response {@link JSONObject} containing Pixabay API response
-     * @param nbCards max number of cards to generate
-     * @throws DatastoreGetter.DataStoreNotAvailableException Exception throwed if {@link DatastoreService} is not available
+     * @param nbCards  max number of cards to generate
      * @return array of {@link Card}
+     * @throws DatastoreGetter.DataStoreNotAvailableException Exception throwed if {@link DatastoreService} is not available
      */
     public static Card[] fromPixabayRespond(JSONObject response, int nbCards) throws DatastoreGetter.DataStoreNotAvailableException {
         String schemasPath = System.getProperty("user.dir") + "/schemas/";
@@ -72,11 +109,11 @@ public class Card {
         }
 
         JSONObject responseJSON = new JSONObject(new JSONTokener(responseIS));
-        Schema schema = SchemaLoader.builder().resolutionScope("file:"+schemasPath).schemaJson(responseJSON).build().load().build();
+        Schema schema = SchemaLoader.builder().resolutionScope("file:" + schemasPath).schemaJson(responseJSON).build().load().build();
 
         try {
             schema.validate(response); // thcardrows a ValidationException if this object is invalid
-        }catch(ValidationException e){
+        } catch (ValidationException e) {
             log.severe("Pixabay api response is not valid : " + e.getMessage());
             e.getCausingExceptions().stream()
                     .map(ValidationException::getMessage)
@@ -87,7 +124,7 @@ public class Card {
         JSONArray images = response.getJSONArray("hits");
         ArrayList<Card> arr = new ArrayList<Card>();
 
-        for (int i=0; i < images.length() && i <= nbCards; i++) {
+        for (int i = 0; i < images.length() && i <= nbCards; i++) {
             JSONObject obj = images.getJSONObject(i);
 
             int id = obj.getInt("id");
@@ -110,12 +147,14 @@ public class Card {
 
     /**
      * Restore Card from an entity obtained from a Datastore query
+     *
      * @param cardEntity {@link Entity} which contain informations of the card
      * @return
      */
-    private static Card restoreFromEntity(Entity cardEntity){
-        if(!cardEntity.getKind().equals("Card"))
+    private static Card restoreFromEntity(Entity cardEntity) {
+        if (!cardEntity.getKind().equals("Card"))
             return null;
+
 
         String tags[] = ((String) cardEntity.getProperty("tags")).split(",\\s*");
         URL pixabayPageURL = null, pixabayImageURL = null, cardImageURL = null;
@@ -134,11 +173,12 @@ public class Card {
 
     /**
      * Restore {@link Card} from store with its id
+     *
      * @param id id of the card
      * @return <ul>
-     *           <li>Return a {@link Card} instance if the id exists in Datastore</li>
-     *           <li>Return null if the id does not exists in Datastore</li>
-     *          </ul>
+     * <li>Return a {@link Card} instance if the id exists in Datastore</li>
+     * <li>Return null if the id does not exists in Datastore</li>
+     * </ul>
      * @throws DatastoreGetter.DataStoreNotAvailableException throwed if the {@link DatastoreService} is not available
      */
     public static Card restoreFromStore(long id) throws DatastoreGetter.DataStoreNotAvailableException {
@@ -159,8 +199,9 @@ public class Card {
 
     /**
      * Restore several cards from the store
+     *
      * @param nbPerPage number of {@link Card}s
-     * @param page page offset
+     * @param page      page offset
      * @return {@link Card}s Array
      * @throws DatastoreGetter.DataStoreNotAvailableException
      */
@@ -170,22 +211,22 @@ public class Card {
         Query query = new Query("Card");
         PreparedQuery preparedQuery = datastore.prepare(query);
 
-        if(nbPerPage <= 0 )
+        if (nbPerPage <= 0)
             nbPerPage = 20;
-        if(page <= 0)
+        if (page <= 0)
             page = 1;
 
         List<Entity> entities;
 
 
-        if(page == 0)
+        if (page == 0)
             entities = preparedQuery.asList(FetchOptions.Builder.withLimit(nbPerPage));
         else
-            entities = preparedQuery.asList(FetchOptions.Builder.withLimit(nbPerPage).offset((nbPerPage)*(page-1)));
+            entities = preparedQuery.asList(FetchOptions.Builder.withLimit(nbPerPage).offset((nbPerPage) * (page - 1)));
 
         Card cards[] = new Card[entities.size()];
         int cpt = 0;
-        for (Entity e: entities) {
+        for (Entity e : entities) {
             cards[cpt] = Card.restoreFromEntity(e);
             cpt++;
         }
@@ -204,6 +245,7 @@ public class Card {
 
     /**
      * Get a random {@link Card} from the store store. (Fortune Wheel method). Use their rarity to draw (tirer in French) a card
+     *
      * @return random Card
      * @throws DatastoreGetter.DataStoreNotAvailableException Exception throwed if {@link DatastoreService} is not available
      */
@@ -215,8 +257,8 @@ public class Card {
         List<Entity> entities = preparedQuery.asList(FetchOptions.Builder.withDefaults());
 
         double maxProba = 0.0;
-        for(Entity e : entities){
-            maxProba += (Double) e.getProperty("probability");
+        for (Entity e : entities) {
+            maxProba += 1 - (Double) e.getProperty("probability");
         }
 
         Random r = new Random();
@@ -225,13 +267,13 @@ public class Card {
         double sum = 0;
         int index = 0;
         long id = -1;
-        while(sum < randomValue && index < entities.size()){
-            sum += (Double) entities.get(index).getProperty("probability");
+        while (sum < randomValue && index < entities.size()) {
+            sum += 1 - (Double) entities.get(index).getProperty("probability");
             id = entities.get(index).getKey().getId();
             index++;
         }
 
-        if(id == -1)
+        if (id == -1)
             return null;
         return Card.restoreFromStore(id);
 
@@ -239,12 +281,12 @@ public class Card {
 
     /**
      * Check if a card is already present in Datastore
-     * @param id of the card
-     * @return  <ul>
-     *              <li>true if the card exists</li>
-     *              <li>false if not</li>
-     *          </ul>
      *
+     * @param id of the card
+     * @return <ul>
+     * <li>true if the card exists</li>
+     * <li>false if not</li>
+     * </ul>
      * @throws DatastoreGetter.DataStoreNotAvailableException Exception throwed if {@link DatastoreService} is not available
      */
     public static boolean existInStore(long id) throws DatastoreGetter.DataStoreNotAvailableException {
@@ -262,40 +304,8 @@ public class Card {
     }
 
     /**
-     * Constructor of {@link Card}
-     * @param id id of the card
-     * @param tags Tags of the card
-     * @param pixabayPageURL URL of the image page
-     * @param pixabayImageURL URL to download Pixabay image
-     * @param pixabayAuthorName Author of the original image
-     * @param probability probability between 0 and 1 to draw the card
-     * @param cardImageURL url of the card image
-     */
-    public Card(long id, String[] tags, URL pixabayPageURL, URL pixabayImageURL, URL cardImageURL, String pixabayAuthorName, double probability) {
-        this(id, tags, pixabayPageURL, pixabayImageURL, pixabayAuthorName);
-        this.probability = probability;
-        this.cardImageURL = cardImageURL;
-    }
-
-    /**
-     * Constructor of {@link Card}
-     * @param id id of the card
-     * @param tags Tags of the card
-     * @param pixabayPageURL URL of the image page
-     * @param pixabayImageURL URL to download Pixabay image
-     * @param pixabayAuthorName Author of the original image
-     */
-    private Card(long id, String[] tags, URL pixabayPageURL, URL pixabayImageURL, String pixabayAuthorName) {
-        this.id = id;
-        this.tags = tags;
-        this.pixabayPageURL = pixabayPageURL;
-        this.pixabayImageURL = pixabayImageURL;
-        this.pixabayAuthorName = pixabayAuthorName;
-        this.probability = Math.random();
-    }
-
-    /**
      * Getter of id
+     *
      * @return id of the card
      */
     public long getId() {
@@ -304,6 +314,7 @@ public class Card {
 
     /**
      * Getter of tags
+     *
      * @return Array of tags ({@link String})
      */
     public String[] getTags() {
@@ -312,6 +323,7 @@ public class Card {
 
     /**
      * Getter of the image page {@link URL}
+     *
      * @return image page {@link URL}
      */
     public URL getPixabayPageURL() {
@@ -320,6 +332,7 @@ public class Card {
 
     /**
      * Getter of the source image {@link URL}
+     *
      * @return source image {@link URL}
      */
     public URL getPixabayImageURL() {
@@ -328,6 +341,7 @@ public class Card {
 
     /**
      * Getter of the image author name
+     *
      * @return name of the author
      */
     public String getPixabayAuthorName() {
@@ -336,23 +350,31 @@ public class Card {
 
     /**
      * Getter of the {@link Card} image URL
+     *
      * @return {@link Card} image URL
      */
-    public URL getCardImageURL() { return cardImageURL; }
+    public URL getCardImageURL() {
+        return cardImageURL;
+    }
 
     /**
      * Getter of the {@link Card} probability
+     *
      * @return {@link Card} probability
      */
-    public double getProbability() { return probability; }
+    public double getProbability() {
+        return probability;
+    }
 
     /**
      * Save an instance of {@link Card} to Store
+     *
      * @throws DatastoreGetter.DataStoreNotAvailableException error throwed if {@link DatastoreService} is not available
      */
     public void saveToStore() throws DatastoreGetter.DataStoreNotAvailableException {
         Entity pixabayImage = new Entity("Card", this.id);
-        pixabayImage.setProperty("tags", new JSONArray(this.tags).join(","));
+
+        pixabayImage.setProperty("tags", String.join(", ", (this.tags)));
         pixabayImage.setProperty("pixabayPageURL", this.pixabayPageURL.toString());
         pixabayImage.setProperty("pixabayImageURL", this.pixabayImageURL.toString());
         pixabayImage.setProperty("pixabayAuthorName", this.pixabayAuthorName);
@@ -365,7 +387,7 @@ public class Card {
 
     private byte[] getBytes(BufferedImage image) throws IOException {
         Iterator iter = ImageIO.getImageWritersByFormatName("png");
-        ImageWriter writer = (ImageWriter)iter.next();
+        ImageWriter writer = (ImageWriter) iter.next();
 
         ByteArrayOutputStream bytes = new ByteArrayOutputStream();
         ImageOutputStream ios = ImageIO.createImageOutputStream(bytes);
@@ -387,7 +409,6 @@ public class Card {
         return bytes.toByteArray();
     }
 
-    // TODO : generate Card image with Graphics2D
     public void generateCardImage() throws IOException, DatastoreGetter.DataStoreNotAvailableException {
         // Making image
         BufferedImage bfImg = new BufferedImage(WIDTH, HEIGHT, BufferedImage.TYPE_3BYTE_BGR);
@@ -399,49 +420,45 @@ public class Card {
         cardGraphics.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY);
 
         Color colorBgText = new Color(254, 212, 149);
-            Color colorBgCard = new Color(164, 107, 20);
+        Color colorBgCard = new Color(164, 107, 20);
         Font fontText = new Font("Monospaced", Font.PLAIN, 15);
 
-        this.addImageAndStarToCard (cardGraphics, colorBgCard);
+        PaintImageUtils.addImageAndStarToCard(this.id, this.pixabayImageURL, this.probability, cardGraphics, colorBgCard);
 
         // Add the rarity text and square
         RoundRectangle2D rect = new RoundRectangle2D.Float(25, 430, 350, 25, 12, 12);
-        addSquareAndTextIntoGraphics2D(cardGraphics,this.deAccent("Rareté : "+ String.format("%.2f", this.probability) +"%"), rect,colorBgText, Color.BLACK, fontText);
+        PaintImageUtils.addSquareAndTextIntoGraphics2D(cardGraphics, PaintImageUtils.deAccent("Rareté : " + String.format("%.2f", this.probability) + "%"), rect, colorBgText, Color.BLACK, fontText);
 
         // Add the tags square and text
         rect = new RoundRectangle2D.Float(25, 460, 350, 25, 12, 12);
-        addSquareAndTextIntoGraphics2D(cardGraphics, "Auteur : " + deAccent(this.pixabayAuthorName), rect,colorBgText, Color.BLACK, fontText);
+        PaintImageUtils.addSquareAndTextIntoGraphics2D(cardGraphics, "Auteur : " + PaintImageUtils.deAccent(this.pixabayAuthorName), rect, colorBgText, Color.BLACK, fontText);
 
         // Add Leyenda square and text
         rect = new RoundRectangle2D.Float(25, 500, 350, 70, 12, 12);
-        addSquareAndTextIntoGraphics2D(cardGraphics, "Tags : " + this.deAccentStringArray(this.tags) , rect,colorBgText, Color.BLACK, fontText);
+        PaintImageUtils.addSquareAndTextIntoGraphics2D(cardGraphics, "Tags : " + PaintImageUtils.deAccentStringArray(this.tags), rect, colorBgText, Color.BLACK, fontText);
 
         // Add the id text
         rect = new RoundRectangle2D.Float(275, 570, 125, 30, 12, 12);
-        addSquareAndTextIntoGraphics2D(cardGraphics, "id : " + this.id, rect,colorBgCard, Color.WHITE, fontText);
+        PaintImageUtils.addSquareAndTextIntoGraphics2D(cardGraphics, "id : " + this.id, rect, colorBgCard, Color.WHITE, fontText);
 
 //        // Add the card url
 //        this.cardImageURL = new URL("https://storage.googleapis.com/cardexchangemaven.appspot.com/"+this.id+".jpg");
-      
+
         GcsFileOptions opt = new GcsFileOptions.Builder().mimeType("image/jpeg").acl("public-read").build();
         GcsService service = GcsServiceFactory.createGcsService();
-        GcsFilename name = new GcsFilename("cardexchangemaven.appspot.com", this.id+".jpg");
+        GcsFilename name = new GcsFilename("cardexchangemaven.appspot.com", this.id + ".jpg");
         Image image = ImagesServiceFactory.makeImage(this.getBytes(bfImg));
 
-        ByteBuffer buff =  ByteBuffer.wrap(image.getImageData());
+        ByteBuffer buff = ByteBuffer.wrap(image.getImageData());
         service.createOrReplace(name, opt, buff);
 
-        if(SystemProperty.environment.value() == SystemProperty.Environment.Value.Production){
+        if (SystemProperty.environment.value() == SystemProperty.Environment.Value.Production) {
             this.cardImageURL = new URL("https://storage.googleapis.com/" + name.getBucketName() + "/" + name.getObjectName());
-        }else{
+        } else {
             ImagesService imagesService = ImagesServiceFactory.getImagesService();
-            this.cardImageURL = new URL(imagesService.getServingUrl(ServingUrlOptions.Builder.withGoogleStorageFileName("/gs/" + name.getBucketName() + "/" + name.getObjectName()) ));
+            this.cardImageURL = new URL(imagesService.getServingUrl(ServingUrlOptions.Builder.withGoogleStorageFileName("/gs/" + name.getBucketName() + "/" + name.getObjectName())));
 
         }
-
-
-
-
 
         this.saveToStore();
     }
@@ -487,245 +504,5 @@ public class Card {
                 '}';
     }
 
-    /**
-     * Return an png img in BufferedImage which can be transform with Graphics2D class
-     * @param imageUrl
-     * @return BufferedImage
-     * @throws IOException
-     */
-    public BufferedImage urlImageToBufferedImage (URL imageUrl) throws IOException, DatastoreGetter.DataStoreNotAvailableException {
-        String tempName = this.getId() + "_temp.png";
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        InputStream inputStream = imageUrl.openStream();
-        int read;
-        while ((read = inputStream.read()) != -1) {
-            baos.write(read);
-        }
-
-        ImagesService imagesService = ImagesServiceFactory.getImagesService();
-        Image image = ImagesServiceFactory.makeImage(baos.toByteArray());
-        // this throws an exception if data is not image or unsupported format
-        // you can wrap this in try..catch and act accordingly
-        try {
-            image.getFormat();
-        } catch (Exception e) {
-            log.severe("Card.java, urlImageToBufferedImage : image format not accepted");
-        }
-        // this is a dummy transform, which do nothing
-        Transform transform = ImagesServiceFactory.makeCrop(0.0, 0.0, 1.0, 1.0);
-        // setting the output to PNG
-        OutputSettings outputSettings = new OutputSettings(ImagesService.OutputEncoding.PNG);
-        outputSettings.setQuality(100);
-        // apply dummy transform and output settings
-        Image newImage = imagesService.applyTransform(transform, image, outputSettings);
-
-
-        // UPLOAD image
-        GcsFileOptions opt = new GcsFileOptions.Builder().mimeType("image/png").acl("public-read").build();
-        GcsService service = GcsServiceFactory.createGcsService();
-        GcsFilename name = new GcsFilename("cardexchangemaven.appspot.com", tempName);
-
-        ByteBuffer buff =  ByteBuffer.wrap(newImage.getImageData());
-        service.createOrReplace(name, opt, buff);
-
-
-        URL pngImg = new URL("https://storage.googleapis.com/cardexchangemaven.appspot.com/"+tempName);
-        BufferedImage newImg = null;
-
-        if(SystemProperty.environment.value() == SystemProperty.Environment.Value.Production){
-            newImg = ImageIO.read(new URL("https://storage.googleapis.com/" + name.getBucketName() + "/" + name.getObjectName()));
-        }else{
-            newImg = ImageIO.read(new URL(imagesService.getServingUrl(ServingUrlOptions.Builder.withGoogleStorageFileName("/gs/" + name.getBucketName() + "/" + name.getObjectName()))));
-        }
-
-
-        service.delete(name);
-
-        return newImg;
-    }
-
-    /**
-     * Draw a String centered in the middle of a y roundedRectangle at 15px to the begin x of the rectangle.
-     *
-     * @param g The Graphics instance.
-     * @param text The String to draw.
-     * @param rect The roundedRectangle to center the text in.
-     */
-    public static void drawCenteredString(Graphics g, String text, RoundRectangle2D rect, Font font) {
-        // Get the FontMetrics
-        FontMetrics metrics = g.getFontMetrics(font);
-        // Determine the X coordinate for the text
-        int x = (int) rect.getX() + 15;
-        // Determine the Y coordinate for the text (note we add the ascent, as in java 2d 0 is top of the screen)
-        int y = (int) (rect.getY() + rect.getHeight() + metrics.getAscent()/2);
-        // Set the font
-        g.setFont(font);
-        // Draw the String
-        g.drawString(text, x, y);
-    }
-
-    public static String deAccentStringArray (String[] strArr) {
-        if(strArr.length == 0)
-            return "";
-
-        String res = strArr[0];
-        for(int i = 1; i < strArr.length; i++){
-            res += ", " + strArr[i];
-        }
-
-        return res;
-    }
-
-    public static String deAccent(String str) {
-//        String nfdNormalizedString = Normalizer.normalize(str, Normalizer.Form.NFD);
-//        Pattern pattern = Pattern.compile("\\p{InCombiningDiacriticalMarks}+");
-//        return pattern.matcher(nfdNormalizedString).replaceAll("");
-
-        String newStr = "";
-
-        for(int i = 0; i < str.length(); i++){
-            char c = str.charAt(i);
-            if(c == 'é'){
-                newStr += 'e';
-            }else if(c =='è'){
-                newStr += 'e';
-            }else if(c =='ê'){
-                newStr += 'e';
-            }else if(c =='à'){
-                newStr += 'a';
-            }else if(c =='â'){
-                newStr += 'a';
-            }else if(c =='ä'){
-                newStr += 'a';
-            }else if(c =='î'){
-                newStr += 'i';
-            }else if(c =='ï'){
-                newStr += 'i';
-            }else if(c =='ô'){
-                newStr += 'o';
-            }else if(c =='ö'){
-                newStr += 'o';
-            }else if(c =='û'){
-                newStr += 'u';
-            }else if(c =='ü'){
-                newStr += 'u';
-            }else{
-                newStr += c;
-            }
-        }
-
-        return newStr;
-    }
-
-    /**
-     * Add a rounded corner to a picture
-     * @param image
-     * @param cornerRadius in pixel
-     * @return
-     */
-    public static BufferedImage makeRoundedCornerImage(BufferedImage image, int cornerRadius) {
-        int w = image.getWidth();
-        int h = image.getHeight();
-        BufferedImage output = new BufferedImage(w, h, BufferedImage.TYPE_INT_ARGB);
-
-        Graphics2D g2 = output.createGraphics();
-
-        // This is what we want, but it only does hard-clipping, i.e. aliasing
-        // g2.setClip(new RoundRectangle2D ...)
-
-        // so instead fake soft-clipping by first drawing the desired clip shape
-        // in fully opaque white with antialiasing enabled...
-        g2.setComposite(AlphaComposite.Src);
-        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        g2.setColor(Color.WHITE);
-        g2.fill(new RoundRectangle2D.Float(0, 0, w, h, cornerRadius, cornerRadius));
-
-        // ... then compositing the image on top,
-        // using the white shape from above as alpha source
-        g2.setComposite(AlphaComposite.SrcAtop);
-        g2.drawImage(image, 0, 0, null);
-
-        g2.dispose();
-
-        return output;
-    }
-
-    /**
-     * Add a square and a text y centering into a graphics2D
-     * @param g
-     * @param text to draw
-     * @param rect rectangle to draw
-     * @param colorBgText color of the rectangle
-     * @param colorText color of the text
-     * @param fontText font of the text
-     */
-    public static void addSquareAndTextIntoGraphics2D(Graphics2D g, String text, RoundRectangle2D rect, Color colorBgText, Color colorText, Font fontText) {
-        g.setColor(colorBgText);
-        g.draw(rect);
-        g.fill(rect);
-        g.setColor(colorText);
-        Card.drawCenteredString(g, text, rect, fontText);
-
-    }
-
-
-    public void addImageAndStarToCard (Graphics2D g, Color colorBgCard) throws IOException, DatastoreGetter.DataStoreNotAvailableException {
-        BufferedImage mig = ImageIO.read(new File("newBack.png"));
-        g.drawImage(mig, 0, 0, 400, 600, null);
-
-        // Get the image of pixabayImageURL
-        BufferedImage newImg = this.urlImageToBufferedImage(this.pixabayImageURL);
-        newImg = makeRoundedCornerImage(newImg, 20);
-        //        fill the card background
-        //g.setColor(colorBgCard);
-        //g.fillRect(0, 0, WIDTH, HEIGHT);
-        // Add image on the top of the card and resize it
-
-
-        int widthFenetre = 350;
-        int heightFenetre = 400;
-        int widthImage = newImg.getWidth();
-        int heightImage = newImg.getHeight();
-
-        double ratioImage = (0.0 + widthImage) / heightImage;
-        double ratioFenetre = (0.0 + widthFenetre) / heightFenetre;
-
-        int newWidth = 0;
-        int newHeight = 0;
-
-        if(ratioFenetre > ratioImage){
-            newWidth = (widthImage * heightFenetre) / heightImage;
-            newHeight = heightFenetre;
-        }else{
-            newWidth = widthFenetre;
-            newHeight = (heightImage * widthFenetre) / widthImage;
-        }
-
-        int top = (heightFenetre - newHeight) / 2;
-        int left = (widthFenetre - newWidth) / 2;
-
-//        g.drawImage(newImg, 25, 25, 350, 400, null);
-
-
-        if(heightImage > widthImage){
-            g.drawImage(newImg, 25 + left, 25, newWidth, newHeight, null);
-        }else{
-            g.drawImage(newImg, 25, 25 + top, newWidth, newHeight, null);
-        }
-
-
-
-        // Get the star point on top right of the card
-//        BufferedImage starImg = ImageIO.read(new URL("https://storage.googleapis.com/cardexchangemaven.appspot.com/_star_point.png"));
-
-        BufferedImage starImg = ImageIO.read(new File("star_point.png"));
-
-        int nbStar = Math.toIntExact(Math.round(10 * this.probability));
-
-        // Add the star to the image
-        for (int i = 0; i < nbStar; i++)
-            g.drawImage(starImg, 350 - (22 * i), 2, 20, 20, null);
-
-    }
 
 }
